refactor: Moved Entities folder to match the server structure
