+++ src/Chrono/ClientApp/src/Entities/Category.ts
