+++ src/Chrono/ClientApp/src/Entities/Note.ts
