+++ src/Chrono/ClientApp/src/Features/TaskLists/TaskListEditControl.less
@@ -66,6 +66,10 @@ div.tasklist {
         > div.extra {
           margin-left: 0 !important;
           margin-right: 0 !important;
+
+          > .label {
+            margin-right: 0.5em !important;
+          }
         }
       }
     }
